Plot new curvature normalization
Plain arctan of curvature is not as smooth as it could be. Try plotting
a new normalization based on derivatives.

This is smoother, but still needs to be scaled. It feels like this is
closer, but I should look at more robust ways of computing the deviation
from G2 continuity, such as matching derivatives directly.