Start to add spiro math
I want to experiment with Euler spirals as the 2-parameter spline
family, and with 4-th order polynomial spirals as the way to do
curvature blending (as opposed to the quintic Hermite). This patch
brings in the spiro computations.

As of this commit, it implements Euler spiral splines with no
curvature blending.