Start implementation of tuner
Instantiates a hacky but verging on usable UI for editing grids of masters.
Right now there's no way to use those, but the pieces are in place.

I'm going to commit this as a checkpoint. There are a few problems. One
that's apparent is that choosing the center point based on G2 continuity
is not very stable. It looks like it will be a lot better choice to
have that as part of the master and interpolate it, which is not G2
(or even G1) continuous by construction, but will be more robust.

Another problem is that the symmetry logic for interpolating beyond the
square is wrong.